Solution:
Given that the array always contains sequential numbers,
this solution identifies the minimum value in order to compare each
value and determine the value's sorted position. This is done by
subtracting the minimum value from the current value in the loop.
If this calculated difference does not equal the current value of i
then a swap occurs where and the counter is incremented. Else, the loop
increments forward to the next value.